Update GcodefilesPanel.vue
Added Scan Metadata option in the gcode file right click menu to call the Metascan Moonraker endpoint to process files that are not auto processed because they are on a network mount or other drive that doesnt support inotify. added a small delay to allow for thumbnail processing as it sometimes took two scans to get the meta and thumbnail.