fix: fix condition in restartServiceNameExists check
Updated the condition in the restartServiceNameExists getter in TheEditor.vue component. Previously, it was returning true if 'restartServiceName' exists. However, this was incorrect as the function should return true only when 'restartServiceName' does not exist. This change fixes an issue where the available services were not correctly checked if 'restartServiceName' existed.

Signed-off-by: Stefan Dej <meteyou@gmail.com>