feat: Add support for hall filament width sensor
The hall_filamend_width_sensor includes a filament runout sensor in
Klipper. This extends the existing filament sensor functionality of
Mainsail to include filament width sensor functionality.

In the UI, the difference between the two is that for a filament width
sensor the filament width will be shown instead of just the string
"detected".